REFACTOR: labels-milestones 공통 component 분리
- list info bar
- search list
