+++ client/src/components/search-list.jsx
@@ -1,0 +1,34 @@
+import React from "react";
+import { BrowserRouter as Router, Link } from "react-router-dom";
+import styled from "styled-components";
+
+import ListInfoBar from "./list-info-bar.jsx";
+import ListSearchResult from "../labels/list-search-result.jsx";
+
+const StyledList = styled.div`
+    width: 900px;
+
+    margin-top: 20px;
+
+    border: 0;
+    border-radius: 5px;
+    box-shadow: 0 0 2px 0 rgb(36, 41, 46);
+
+    font-family: "Noto Sans KR", sans-serif;
+    font-weight: 400;
+    font-size: 12px;
+`;
+
+const labelsData = JSON.parse(localStorage.getItem("labelsData"));
+const numOfLabel = labelsData?.length || 0;
+
+const SearchList = () => {
+    return (
+        <StyledList>
+            <ListInfoBar numOfLabel={numOfLabel} />
+            <ListSearchResult labelsData={labelsData} numOfLabel={numOfLabel} />
+        </StyledList>
+    );
+};
+
+export default SearchList;
